QRC-98 (#53)
* QRC-98
Сделал логику по бронированию билетов и опций
Сделал логику по снятию бронирования с билетов и опций

* QRC-98
Добавлено:
1. фикс авторизации
2. описание в swagger

* QRC-98
fix notes

* QRC-98

* QRC-98

* QRC-98

* QRC-98
smallFix

+++ src/main/java/ru/softdarom/qrcheck/events/model/dto/OptionDto.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.Generated;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Data
 @Generated
@@ -28,8 +29,8 @@ public class OptionDto {
 
     @NotNull
     @JsonProperty("cost")
-    private Double cost;
+    private BigDecimal cost;
 
     @JsonProperty("price")
-    private Double price;
+    private BigDecimal price;
 }
